+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 Phormium Changelog
 ==================
 
+0.2 / 2013-05-10
+----------------
+
+* Added transactions
+* Added QuerySet::dump()
+* Added logging via [Apache log4php](http://logging.apache.org/log4php/)
+* Added composite filters
+
 0.1 / 2013-04-25
 ----------------
 
-* first release
+* Initial release